chore(interpreter_monty): satisfy DCM warnings on touched files
Reduces `dcm analyze` to zero issues. All targets are pre-existing
or surfaced during the 0.17.1 port:

  - Extract `ToolNameMapping` into its own file so
    `tool_definition_converter.dart` is a pure-functions module —
    fixes `prefer-match-file-name`.
  - Replace `result.error!` and `result.value.dartValue! as Map`
    in `SchemaExecutor.validate` with explicit null + type checks;
    a non-Map return now throws `StateError` with a descriptive
    message — fixes both `avoid-non-null-assertion` warnings.
  - Switch the JSON-Schema `required` field cast from `List<dynamic>`
    to `List<Object?>` — fixes the `avoid-dynamic` warning while
    keeping the `e.toString()` mapping intact.
  - Add missing blank lines before `return` in 4 spots in
    `_toPythonLiteral` and `_jsonSchemaTypeToParamType` —
    fixes `newline-before-return`.
  - Convert 7 `params[0]` / `mappings[0]` accessors in the converter
    test to `.first` — fixes `prefer-first`.

Net DCM result for this package: 16 issues → 0.
